move RestGatewayClient to rest_gateway.rs
